requirements.txt
fastapi
uvicorn
requests
python-dotenv
ffmpeg-python
Pillow
celery
redis
